% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3s.R
\name{M3S}
\alias{M3S}
\title{M3S}
\usage{
M3S(x, sampling = 100, significant.cut = 0.1,
  verbose = interactive())
}
\arguments{
\item{x}{A matrix or an array stores a the data, if a matrix, each row is one feature and each column is one gene}

\item{sampling}{Unused for now}

\item{significant.cut}{Cut off for significant.zeros}

\item{verbose}{display promote or not}
}
\value{
statistical indicator for available modes
}
\description{
M3S
}
