% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.R
\name{M3Sfit}
\alias{M3Sfit}
\title{M3S Fit}
\usage{
M3Sfit(x, normalization = c("none", "logplus", "log", "cpm.none",
  "cpm.logplus", "cpm.log"), distribution = c("G", "MG", "LTMG", "LTG"))
}
\arguments{
\item{x}{}

\item{distribution}{models}
}
\description{
M3S Fit
}
